requirements.txt
numpy==1.23.5
pandas==1.5.3
matplotlib==3.7.0
seaborn==0.12.2
streamlit==1.20.0
plotly==5.13.1
folium==0.14.0
Pillow==9.4.0
streamlit-folium==0.15.0